+++ server/src/main/java/com/hedera/block/server/mediator/PollerImpl.java
@@ -2,7 +2,6 @@
 package com.hedera.block.server.mediator;
 
 import com.lmax.disruptor.EventPoller;
-import com.lmax.disruptor.RingBuffer;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Objects;
 
@@ -10,23 +9,19 @@ public class PollerImpl<V> implements Poller<V> {
 
     private final EventPoller<V> poller;
     private final BatchedData<V> polledData;
-    private final RingBuffer<V> ringBuffer;
 
     /**
      * Leverage the provided ring buffer and batch size to poll for events.
      *
-     * @param ringBuffer the ring buffer to poll events from
+     * @param eventPoller the event poller
      * @param batchSize the size of the batch
      */
-    PollerImpl(@NonNull final RingBuffer<V> ringBuffer, final int batchSize) {
-        this.ringBuffer = Objects.requireNonNull(ringBuffer);
-        this.poller = ringBuffer.newPoller();
-        ringBuffer.addGatingSequences(poller.getSequence());
+    PollerImpl(@NonNull final EventPoller<V> eventPoller, final int batchSize) {
+        this.poller = Objects.requireNonNull(eventPoller);
         this.polledData = new BatchedData<>(batchSize);
     }
 
     @Override
-    @NonNull
     public V poll() throws Exception {
         if (polledData.getMsgCount() > 0) {
             return polledData.pollMessage();
@@ -34,11 +29,6 @@ public class PollerImpl<V> implements Poller<V> {
 
         loadNextValues(poller, polledData);
         return polledData.getMsgCount() > 0 ? polledData.pollMessage() : null;
-    }
-
-    @Override
-    public void unsubscribe() {
-        ringBuffer.removeGatingSequence(poller.getSequence());
     }
 
     private EventPoller.PollState loadNextValues(final EventPoller<V> poller, final BatchedData<V> batch)
