+++ simulator/src/main/resources/app.properties
@@ -33,7 +33,7 @@ generator.managerImplementation=BlockAsFileLargeDataSets
 
 #blockStream.maxBlockItemsToStream=100_000_000
 blockStream.streamingMode=MILLIS_PER_BLOCK
-blockStream.millisecondsPerBlock=500
+blockStream.millisecondsPerBlock=1000
 #blockStream.blockItemsBatchSize=1_000
 #blockStream.delayBetweenBlockItems=3_000_000
